Split off stuff from CameraStageBaseScreen that isn't the class itself
